fix(sprint): production readiness - security, cart, checkout, prices
Teams 3/7/8/2 frontend:
- cart.ts: fix missing Bearer prefix on all 6 auth headers
- lib/authHeaders.ts: new centralized getAuthHeader() utility
- cache.ts: wrap debug console.log in __DEV__ guard
- checkout-v2.tsx: replace token polling loop with getToken({forcedRefresh})
- checkout-v2.tsx: move resetCart() out of pre-redirect — success/failure pages handle it
- schemas/order.py: fix OrderItemResponse.price int -> float (was truncating cents)
- cartStore.ts: on error rollback call syncWithServer() instead of stale client snapshot

Teams 6 security/config:
- ci.yml: add Gitleaks secret-scan and Bandit SAST parallel jobs
- config.py: add SENTRY_DSN field, _validate_production_config() for production startup
- scripts/smoke_tests.py: post-deploy endpoint health checks

+++ backend/app/schemas/order.py
@@ -68,7 +68,7 @@ class OrderItemResponse(BaseModel):
     order_id: int
     product_id: int
     quantity: int
-    price: int
+    price: float  # DB stores Numeric(10,2); float preserves cents without breaking JSON serialization
 
     model_config = ConfigDict(from_attributes=True)
